+++ src/extension.ts
@@ -49,20 +49,24 @@ export function activate(context: ExtensionContext) {
   Utils.init(context);
 
   let logtalkCommands = [
-    { command: "logtalk.load.directory",         callback: uri  => LogtalkTerminal.loadDirectory(uri, linter)},
-    { command: "logtalk.load.file",              callback: uri  => LogtalkTerminal.loadFile(uri, linter)},
-    { command: "logtalk.make.reload",            callback: async (uri)  => LogtalkTerminal.makeReload(uri)},
-    { command: "logtalk.make.check",             callback: async (uri)  => LogtalkTerminal.makeCheck(uri)},
-    { command: "logtalk.run.tests",              callback: uri  => LogtalkTerminal.runTests(uri)},
-    { command: "logtalk.run.doclet",             callback: uri  => LogtalkTerminal.runDoclet(uri)},
-    { command: "logtalk.scan.deadCode",          callback: uri  => LogtalkTerminal.scanForDeadCode(uri, deadCodeScanner)},
-    { command: "logtalk.generate.documentation", callback: uri  => LogtalkTerminal.genDocumentation(uri, documentationLinter)},
-    { command: "logtalk.generate.diagrams",      callback: uri  => LogtalkTerminal.genDiagrams(uri)},
-    { command: "logtalk.open",                   callback: ()   => LogtalkTerminal.openLogtalk()},
-    { command: "logtalk.rscan.deadCode",         callback: uri  => LogtalkTerminal.rscanForDeadCode(uri, deadCodeScanner)},
-    { command: "logtalk.run.testers",            callback: uri  => LogtalkTerminal.runTesters(uri)},
-    { command: "logtalk.run.doclets",            callback: uri  => LogtalkTerminal.runDoclets(uri)},
-    { command: "logtalk.open.parentFile",        callback: uri  => LogtalkTerminal.openParentFile(uri)}
+    // workspace commands
+    { command: "logtalk.open",                    callback: ()   => LogtalkTerminal.openLogtalk()},
+    { command: "logtalk.rscan.deadCode",          callback: uri  => LogtalkTerminal.rscanForDeadCode(uri, deadCodeScanner)},
+    { command: "logtalk.rgenerate.documentation", callback: uri  => LogtalkTerminal.rgenDocumentation(uri, documentationLinter)},
+    { command: "logtalk.rgenerate.diagrams",      callback: uri  => LogtalkTerminal.rgenDiagrams(uri)},
+    { command: "logtalk.run.testers",             callback: uri  => LogtalkTerminal.runTesters(uri)},
+    { command: "logtalk.run.doclets",             callback: uri  => LogtalkTerminal.runDoclets(uri)},
+    // directory and file commands
+    { command: "logtalk.load.directory",          callback: uri  => LogtalkTerminal.loadDirectory(uri, linter)},
+    { command: "logtalk.load.file",               callback: uri  => LogtalkTerminal.loadFile(uri, linter)},
+    { command: "logtalk.make.reload",             callback: async (uri)  => LogtalkTerminal.makeReload(uri)},
+    { command: "logtalk.make.check",              callback: async (uri)  => LogtalkTerminal.makeCheck(uri)},
+    { command: "logtalk.run.tests",               callback: uri  => LogtalkTerminal.runTests(uri)},
+    { command: "logtalk.run.doclet",              callback: uri  => LogtalkTerminal.runDoclet(uri)},
+    { command: "logtalk.scan.deadCode",           callback: uri  => LogtalkTerminal.scanForDeadCode(uri, deadCodeScanner)},
+    { command: "logtalk.generate.documentation",  callback: uri  => LogtalkTerminal.genDocumentation(uri, documentationLinter)},
+    { command: "logtalk.generate.diagrams",       callback: uri  => LogtalkTerminal.genDiagrams(uri)},
+    { command: "logtalk.open.parentFile",         callback: uri  => LogtalkTerminal.openParentFile(uri)}
   ];
 
   logtalkCommands.map(command => {